refactor: Move request and response classes to dto package for improved structure
